fix not selecting users from friendadapter in chooseusersactivity

+++ app/src/main/kotlin/me/alexeyterekhov/vkfilter/GUI/ChooseUsersActivity/UserList/FriendListAdapter.kt
@@ -63,14 +63,14 @@ public class FriendListAdapter(
 
             name setText TextFormat.userTitle(user, false)
 
-            val id = user.id
+            val id = user.id.toLong()
             with (checkBox) {
                 setOnCheckedChangeListener(null)
                 setChecked(selected contains id)
                 setOnCheckedChangeListener {
                     view, checked ->
                     if (checked) {
-                        selected add id.toLong()
+                        selected add id
                     } else {
                         selected remove id
                     }
